refactor(updateNoticeModal): move updateNoticeModal to src directory
- Relocate updateNoticeModal.ts to src directory
- Maintain existing functionality without changes
- Ensure consistent formatting and styling
- Update file structure for better organization
